backend/requirements.txt
django-ninja

